QVM changes in Wolf games to be able to use 64 bit dlls.

+++ source/enemy_territory/src/win32/win_main.c
@@ -564,8 +564,8 @@ char* Sys_GetDLLName( const char *name ) {
 // fqpath param added 2/15/02 by T.Ray - Sys_LoadDll is only called in vm.c at this time
 // fqpath will be empty if dll not loaded, otherwise will hold fully qualified path of dll module loaded
 // fqpath buffersize must be at least MAX_QPATH+1 bytes long
-void * QDECL Sys_LoadDll( const char *name, char *fqpath, int( QDECL **entryPoint ) ( int, ... ),
-						  int ( QDECL *systemcalls )( int, ... ) ) {
+void * QDECL Sys_LoadDll( const char *name, char *fqpath, intptr_t( QDECL **entryPoint ) ( int, ... ),
+						  intptr_t ( QDECL *systemcalls )( int, ... ) ) {
 	HINSTANCE libHandle;
 	void ( QDECL * dllEntry )( int ( QDECL *syscallptr )( int, ... ) );
 	char    *basepath;
